@@ -3,8 +3,9 @@
     "ext-simplexml": "*",
     "ext-dom": "*",
     "ext-libxml": "*",
-    "php": ">=7.2.0",
-    "symfony/console": "^5.1"
+    "php": ">=7.4.0",
+    "symfony/console": "^5.1",
+    "php-di/php-di": "^6.4"
   },
   "require-dev": {
     "roave/security-advisories": "dev-master",
