@@ -2,112 +2,81 @@
 
 namespace Life;
 
+use Life\Helper\XmlFileWriter;
+use Life\Service\GameService;
+
 class Game
 {
-    /** @var int */
-    private $iterationsCount;
 
-    /** @var int */
-    private $size;
+    private GameService $gameService;
 
-    /** @var int */
-    private $species;
+    private int $iterationsCount;
+
+    private int $size;
+
+    private int $species;
 
     /**
-     * @var int[][]|null[][]
      * Array of available cells in the game with size x size dimensions
-     * Indexed by y coordinate and than x coordinate
+     * Indexed by y coordinate and then x coordinate
+     *
+     * @var int[][]|null[][]
      */
-    private $cells;
+    private array $cells;
 
-    public function run(string $inputFile, string $outputFile): void
-    {
-        $input = new XmlFileReader($inputFile);
-        $output = new XmlFileWriter($outputFile);
+    private XmlFileWriter $output;
 
-        [$size, $species, $cells, $iterationsCount] = $input->loadFile();
-
+    /**
+     * @param GameService $gameService
+     * @param int $size
+     * @param int $species
+     * @param int[][]|null[][] $cells
+     * @param int $iterationsCount
+     * @param XmlFileWriter $output
+     */
+    public function __construct(
+        GameService $gameService,
+        int $size,
+        int $species,
+        array $cells,
+        int $iterationsCount,
+        XmlFileWriter $output
+    ) {
+        $this->gameService = $gameService;
         $this->size = $size;
         $this->species = $species;
         $this->cells = $cells;
         $this->iterationsCount = $iterationsCount;
-
-        for ($i = 0; $i < $this->iterationsCount; $i++) {
-            $newCells = [];
-            for ($y = 0; $y < $this->size; $y++) {
-                $newCells[] = [];
-                for ($x = 0; $x < $this->size; $x++) {
-                    $newCells[$y][$x] = $this->evolveCell($x, $y);
-                }
-            }
-            $this->cells = $newCells;
-        }
-
-        $output->saveWorld($this->size, $this->species, $this->cells);
+        $this->output = $output;
     }
 
-    private function evolveCell(int $x, int $y): ?int
+    public function run(): void
     {
-        $cell = $this->cells[$y][$x];
-        $neighbours = [];
-
-        if ($y - 1 >= 0 && $x - 1 >= 0) {
-            $neighbours[] = $this->cells[$y - 1][$x - 1];
-        }
-        if ($y - 1 >= 0) {
-            $neighbours[] = $this->cells[$y - 1][$x];
-        }
-        if ($y - 1 >= 0 && $x + 1 < $this->size) {
-            $neighbours[] = $this->cells[$y - 1][$x + 1];
+        $iteratedCells = [];
+        for ($i = 0; $i < $this->iterationsCount; $i++) {
+            $iteratedCells = $this->iterateEvolution();
         }
 
-        if ($x - 1 >= 0) {
-            $neighbours[] = $this->cells[$y][$x - 1];
-        }
-        if ($x + 1 < $this->size) {
-            $neighbours[] = $this->cells[$y][$x + 1];
-        }
+        $this->output->saveWorld($this->size, $this->species, $iteratedCells);
+    }
 
-        if ($y + 1 < $this->size && $x - 1 >= 0) {
-            $neighbours[] = $this->cells[$y + 1][$x - 1];
-        }
-        if ($y + 1 < $this->size) {
-            $neighbours[] = $this->cells[$y + 1][$x];
-        }
-        if ($y + 1 < $this->size && $x + 1 < $this->size) {
-            $neighbours[] = $this->cells[$y + 1][$x + 1];
-        }
-
-        $sameSpeciesCount = 0;
-        foreach ($neighbours as $neighbour) {
-            if ($neighbour === $cell) {
-                $sameSpeciesCount++;
+    /**
+     * @return int[][]|null[][]
+     */
+    private function iterateEvolution(): array
+    {
+        $iteratedCells = [];
+        for ($y = 0; $y < $this->size; $y++) {
+            for ($x = 0; $x < $this->size; $x++) {
+                $iteratedCells[$y][$x] = $this->gameService->evolveCell(
+                    $this->cells,
+                    $this->size,
+                    $this->species,
+                    $x,
+                    $y
+                );
             }
         }
-
-        if ($cell !== null && $sameSpeciesCount >= 2 && $sameSpeciesCount <= 3) {
-            return $cell;
-        }
-
-        $speciesForBirth = [];
-        for ($i = 0; $i < $this->species; $i++) {
-            $oneSpeciesCount = 0;
-
-            foreach ($neighbours as $neighbour) {
-                if ($neighbour === $i) {
-                    $oneSpeciesCount++;
-                }
-            }
-
-            if ($oneSpeciesCount === 3) {
-                $speciesForBirth[] = $i;
-            }
-        }
-
-        if (count($speciesForBirth) > 0) {
-            return $speciesForBirth[array_rand($speciesForBirth)];
-        }
-
-        return null;
+        return $iteratedCells;
     }
 }
