@@ -1,10 +1,13 @@
 <?php
 
+use DI\Container;
 use Life\RunGameCommand;
 use Symfony\Component\Console\Application;
 
 require './vendor/autoload.php';
 
 $app = new Application();
-$app->add(new RunGameCommand());
+$container = new Container();
+$command = $container->get(RunGameCommand::class);
+$app->add($command);
 $app->run();
